deeplearn-classroom/backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
tensorflow>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
mysql-connector-python>=8.1.0
gunicorn>=21.2.0
mediapipe==0.10.9
opencv-python>=4.8.0
moviepy
openai-whisper
SpeechRecognition
Pillow
ffmpeg-python
